Move interpreter files to interpreter subdir
