+++ src/main.cpp
@@ -18,8 +18,6 @@
 #include "DebugReport.h"
 #include "ErrorReport.h"
 #include "Global.h"
-#include "InterpreterEngine.h"
-#include "InterpreterProgInterface.h"
 #include "ParserDriver.h"
 #include "Synthesiser.h"
 #include "ast/AstNode.h"
@@ -35,6 +33,8 @@
 #include "ast/transform/IOAttributesTransformer.h"
 #include "ast/transform/IODefaultsTransformer.h"
 #include "config.h"
+#include "interpreter/InterpreterEngine.h"
+#include "interpreter/InterpreterProgInterface.h"
 #include "ram/RamNode.h"
 #include "ram/RamProgram.h"
 #include "ram/RamTranslationUnit.h"
