+++ src/tests/ram_relation_test.cpp
@@ -19,8 +19,8 @@
 #include "DebugReport.h"
 #include "ErrorReport.h"
 #include "Global.h"
-#include "InterpreterEngine.h"
 #include "RelationTag.h"
+#include "interpreter/InterpreterEngine.h"
 #include "ram/RamExpression.h"
 #include "ram/RamOperation.h"
 #include "ram/RamProgram.h"
